Move create_well_acquisition_dict & _split_well_path_image_path into fractal_tasks_core.utils
