+++ fractal_tasks_core/tasks/init_group_by_well_for_multiplexing.py
@@ -16,7 +16,7 @@ import logging
 
 from pydantic import validate_call
 
-from fractal_tasks_core.tasks._registration_utils import (
+from fractal_tasks_core.utils import (
     create_well_acquisition_dict,
 )
